Changes following comments on PR
