+++ templates/includes/article-spotlight.html
@@ -4,7 +4,7 @@
     <h2 class="extra-padding">{% translate title %}</h2>
     <div class="featured-article">
         {% if page.hero_image %}
-            <a href="{% pageurl page %}" class="featured-article__image-link" tabindex="-1" aria-hidden="true" data-component-name="Featured Story: {{ page.title }}" data-link-type="Image" {% if page.newly_published_at > page.new_label_end_date %}data-label="New"{% endif %}>
+            <a href="{% pageurl page %}" class="featured-article__image-link" tabindex="-1" aria-hidden="true" data-component-name="Featured Story: {{ page.title }}" data-link-type="Image" {% if page.is_new %}data-label="New"{% endif %}>
                 <picture>
                     {% image page.teaser_image fill-500x400-c100 as hero %}
                     <source media="(max-width: 768px)" srcset="{{ hero.url }}"/>
@@ -22,7 +22,7 @@
             <h2 class="featured-article__heading">
                 <small>{% trans "Spotlight on" %}</small>
                 {{ page.title }}
-                {% if page.newly_published_at > page.new_label_end_date %}
+                {% if page.is_new %}
                     <span aria-describedby="article-desc" role="link">NEW</span>
                 {% endif %}
             </h2>
@@ -30,7 +30,7 @@
             <p id="article-desc" class="aria-desc">This is a link to a new story about {{ page.title }}</p>
 
             <p>{{ page.teaser_text }}</p>
-            <a class="tna-button--dark" href="{% pageurl page %}" data-component-name="Featured Story: {{ page.title }}" data-link-type="Button" {% if page.newly_published_at > page.new_label_end_date %}data-label="New"{% endif %}>Read</a>
+            <a class="tna-button--dark" href="{% pageurl page %}" data-component-name="Featured Story: {{ page.title }}" data-link-type="Button" {% if page.is_new %}data-label="New"{% endif %}>Read</a>
         </div>
     </div>
 </div>
