+++ templates/includes/card-group-secondary-nav.html
@@ -16,7 +16,7 @@
             {% endif %}
             data-link-type="Card image"
             data-card-title="{{ page.title }}"
-            {% if page.newly_published_at > page.new_label_end_date %}
+            {% if page.is_new %}
                 data-label = "New"
             {%endif%}
             aria-hidden="true"
@@ -45,11 +45,11 @@
                                 {% endif %}
                                 data-link-type="Card text"
                                 data-card-title="{{ page.title }}"
-                                {% if page.newly_published_at > page.new_label_end_date %}
+                                {% if page.is_new %}
                             data-label = "New"
                             {%endif%}
                             >{{ page.title }}</a>
-                        {% if page.newly_published_at > page.new_label_end_date %}
+                        {% if page.is_new %}
                         <span>NEW</span>
 
                         {% endif %}
